+++ src/netedit/changes/GNEChange_MeanData.cpp
@@ -24,6 +24,9 @@
 #include <config.h>
 
 #include <netedit/GNENet.h>
+#include <netedit/GNEViewNet.h>
+#include <netedit/GNEViewParent.h>
+#include <netedit/GNEApplicationWindow.h>
 #include <netedit/elements/data/GNEMeanData.h>
 
 #include "GNEChange_MeanData.h"
@@ -45,17 +48,20 @@ GNEChange_MeanData::GNEChange_MeanData(GNEMeanData* meanData, bool forward) :
 
 
 GNEChange_MeanData::~GNEChange_MeanData() {
-    myMeanData->decRef("GNEChange_MeanData");
-    if (myMeanData->unreferenced()) {
-        // show extra information for tests
-        WRITE_DEBUG("Deleting unreferenced " + myMeanData->getTagStr());
-        // make sure that MeanData isn't in net before removing
-        if (myMeanData->getNet()->getAttributeCarriers()->retrieveMeanData(myMeanData->getTagProperty().getTag(), myMeanData->getID(), false)) {
-            // delete MeanData from net
-            myMeanData->getNet()->getAttributeCarriers()->deleteMeanData(myMeanData);
+    // only continue we have undo-redo mode enabled
+    if (myMeanData->getNet()->getViewNet()->getViewParent()->getGNEAppWindows()->isUndoRedoAllowed()) {
+        myMeanData->decRef("GNEChange_MeanData");
+        if (myMeanData->unreferenced()) {
+            // show extra information for tests
+            WRITE_DEBUG("Deleting unreferenced " + myMeanData->getTagStr());
+            // make sure that MeanData isn't in net before removing
+            if (myMeanData->getNet()->getAttributeCarriers()->retrieveMeanData(myMeanData->getTagProperty().getTag(), myMeanData->getID(), false)) {
+                // delete MeanData from net
+                myMeanData->getNet()->getAttributeCarriers()->deleteMeanData(myMeanData);
+            }
+            // delete mean data
+            delete myMeanData;
         }
-        // delete mean data
-        delete myMeanData;
     }
 }
